requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
cuda-python==12.5.0
cudf-cu12==24.2.2
cuml-cu12==24.2.0
cupy-cuda12x==13.2.0
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
dask==2024.1.1
dask-cuda==24.2.0
dask-cudf-cu12==24.2.2
distributed==2024.1.1
fastrlock==0.8.2
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.25.0
fsspec==2024.6.1
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.0
h5py==3.11.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
Jinja2==3.1.4
joblib @ file:///croot/joblib_1718217211762/work
keras==3.4.1
kiwisolver @ file:///opt/conda/conda-bld/kiwisolver_1653292039266/work
libclang==18.1.1
llvmlite==0.43.0
locket==1.0.0
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1647441664166/work
mdurl==0.1.2
mkl-fft==1.3.0
mkl-random==1.0.2
mkl-service==2.3.0
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.0.8
munkres==1.1.4
namex==0.0.8
networkx==3.2.1
nibabel==5.2.1
nilearn==0.10.4
numba==0.60.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1607448141537/work
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
nvtx==0.2.10
opt-einsum==3.3.0
optree==0.12.1
packaging @ file:///croot/packaging_1720101850331/work
pandas==1.5.3
partd==1.4.2
Pillow==9.0.1
prettytable==3.12.0
protobuf==4.25.3
psutil==6.0.0
pyarrow==14.0.2
Pygments==2.18.0
pylibraft-cu12==24.2.0
pynvjitlink-cu12==0.3.0
pynvml==11.4.1
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytz==2024.1
PyYAML==6.0.1
raft-dask-cu12==24.2.0
rapids-dask-dependency==24.2.0
requests==2.32.3
rich==13.7.1
rmm-cu12==24.2.0
safetensors==0.4.3
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1642617106979/work
scipy==1.13.1
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
sortedcontainers==2.4.0
sympy==1.13.0
tblib==3.0.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl @ file:///croot/threadpoolctl_1719407800858/work
timm==1.0.7
toolz==0.12.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tornado @ file:///tmp/build/80754af9/tornado_1606942317143/work
tqdm==4.66.4
treelite==4.0.0
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
ucx-py-cu12==0.36.0
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
xgboost==1.7.3
zict==3.0.0
zipp==3.19.2
